a2a_agent/pyproject.toml
[project]
name = "a2a-sitecheck"
version = "0.1.0"
description = "A2A Agent for Site Check"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.25",
    "fastapi>=0.135.2",
    "uvicorn>=0.42.0",
    "mcp>=0.9.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "openai>=2.30.0",
]

[tool.uv]
package = false
